+++ app/helpers/application_helper.rb
@@ -55,6 +55,10 @@ module ApplicationHelper
     target_link_as_blank(markdown.render(replace_numbers_with_ascii(text.to_s)))
   end
 
+  def beta_enabled?(current_user)
+    current_user && current_user.beta_enabled?
+  end
+
   private
 
     def target_link_as_blank(text)