+++ db/migrate/20150327161519_add_beta_enabled_to_users.rb
@@ -1,0 +1,5 @@
+class AddBetaEnabledToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :beta_enabled, :boolean, null: false, default: false
+  end
+end